infra/terraform/modules/gke-gpu/functions/requirements.txt
functions-framework==3.*
google-cloud-container==2.*
flask==3.*
